refactor(condo): DOMA-4226 removed 'common' module from 'sbbol' in favour of 'constants' and 'utils`
