+++ apps/condo/domains/organization/integrations/sbbol/utils/getSbbolUserInfoErrors.js
@@ -1,0 +1,12 @@
+const Ajv = require('ajv')
+const { SbbolUserInfoSchema } = require('../constants')
+
+function getSbbolUserInfoErrors (userInfo) {
+    const ajv = new Ajv()
+    ajv.validate(SbbolUserInfoSchema, userInfo)
+    return (ajv.errors) ? ajv.errors.map(x => x.message) : []
+}
+
+module.exports = {
+    getSbbolUserInfoErrors,
+}
