+++ apps/condo/domains/organization/integrations/sbbol/utils/index.js
@@ -1,10 +1,12 @@
 const { buildBicryptId } = require('./buildBicryptId')
 const { changeClientSecret } = require('./changeClientSecret')
 const { getOrganizationAccessToken, initializeSbbolAuthApi } = require('./getOrganizationAccessToken')
+const { getSbbolUserInfoErrors } = require('./getSbbolUserInfoErrors')
 
 module.exports = {
     buildBicryptId,
     changeClientSecret,
     getOrganizationAccessToken,
     initializeSbbolAuthApi,
+    getSbbolUserInfoErrors,
 }
